feat(AgmCoreModule): support loader config
Closes #609
